Updated dates for 2026 conferences
Just came across this repo and this is the best resource I've found for information of all relevant ML conferences! I noticed some of the entries were out of date so I went ahead and made some updates.

I updated all conferences that have already passed as of October 5, 2025. I could not find any information about if there will be a COLING conference in 2026 so I left it as is.